pyproject.toml
[project]
name = "ai-meeting-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.9.0",
    "groq>=0.37.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-groq>=1.1.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "torch>=2.10.0",
    "transformers>=5.3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
